+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Add `AutoSuggestBox.maxPopupHeight` ([#677](https://github.com/bdlukaa/fluent_ui/issues/677))
+
 ## 4.1.4
 
 - Avoid overflow in `DatePicker` and `TimePicker` popup ([#663](https://github.com/bdlukaa/fluent_ui/issues/663))